+++ src/store/be_readable_pair_/index.js
@@ -2,15 +2,15 @@ import { be_ } from '@ctx-core/object'
 /** @typedef {import('@ctx-core/object').be__params_T}be__params_T */
 /** @typedef {import('@ctx-core/object').Ctx}Ctx */
 /** @typedef {import('svelte/store').Readable}Readable */
-/** @typedef {import('./index.d.ts').be_derived_pair_T}be_derived_pair_T */
+/** @typedef {import('./index.d.ts').be_readable_pair_T}be_readable_pair_T */
 /**
  * @param {string|null|undefined|((ctx:Ctx)=>Readable)}id_OR_derived__new
  * @param {((ctx:Ctx)=>Readable)|be__params_T}[derived__new_OR_be__params]
  * @param {be__params_T}[be__params]
- * @returns {be_derived_pair_T}
+ * @returns {be_readable_pair_T}
  * @private
  */
-export function be_derived_pair_(
+export function be_readable_pair_(
 	id_OR_derived__new,
 	derived__new_OR_be__params,
 	be__params
@@ -26,7 +26,7 @@ export function be_derived_pair_(
 		id = id_OR_derived__new
 		derived__new = derived__new_OR_be__params
 	}
-	if (!derived__new) throw new Error('be_derived_pair_|derived__new argument is required')
+	if (!derived__new) throw new Error('be_readable_pair_|derived__new argument is required')
 	const _be_ =
 		(be__params && be__params.be_)
 		?? be_
@@ -41,4 +41,7 @@ export function be_derived_pair_(
 		val__,
 		val_,
 	]
+}
+export {
+	be_readable_pair_ as be_derived_pair_,
 }