+++ src/store/be_readable_pair_/index.test.ts
@@ -1,8 +1,8 @@
 import { be_, ctx_, type MapCtx } from '@ctx-core/object'
 import { test } from 'uvu'
 import { equal } from 'uvu/assert'
-import { writable_, be_derived_pair_, derived_ } from '../index.js'
-test('be_derived_pair_|+base_name|+derived__new|+be__params', ()=>{
+import { writable_, be_readable_pair_, derived_ } from '../index.js'
+test('be_readable_pair_|+base_name|+derived__new|+be__params', ()=>{
 	const base_ = writable_(1)
 	const ctx = ctx_()
 	const is_source_ = (map_ctx:MapCtx)=>map_ctx === ctx
@@ -14,7 +14,7 @@ test('be_derived_pair_|+base_name|+derived__new|+be__params', ()=>{
 	const [
 		foobar__,
 		foobar_,
-	] = be_derived_pair_(
+	] = be_readable_pair_(
 		'foobar',
 		()=>
 			derived_(base_, base=>
@@ -30,12 +30,12 @@ test('be_derived_pair_|+base_name|+derived__new|+be__params', ()=>{
 	equal(foobar__([ctx_(), ctx]).$, 3)
 	equal(foobar_([ctx_(), ctx]), 3)
 })
-test('be_derived_pair_|+base_name|+derived__new|-be__params', ()=>{
+test('be_readable_pair_|+base_name|+derived__new|-be__params', ()=>{
 	const base_ = writable_(1)
 	const [
 		foobar__,
 		foobar_,
-	] = be_derived_pair_(
+	] = be_readable_pair_(
 		'foobar',
 		()=>
 			derived_(base_, base=>
@@ -47,7 +47,7 @@ test('be_derived_pair_|+base_name|+derived__new|-be__params', ()=>{
 	equal(foobar__(ctx).$, 3)
 	equal(foobar_(ctx), 3)
 })
-test('be_derived_pair_|-base_name|+derived__new|+be__params', ()=>{
+test('be_readable_pair_|-base_name|+derived__new|+be__params', ()=>{
 	const base_ = writable_(1)
 	const ctx = ctx_()
 	const is_source_ = (map_ctx:MapCtx)=>map_ctx === ctx
@@ -59,7 +59,7 @@ test('be_derived_pair_|-base_name|+derived__new|+be__params', ()=>{
 	const [
 		foobar__,
 		foobar_,
-	] = be_derived_pair_(
+	] = be_readable_pair_(
 		()=>
 			derived_(base_,
 				base=>base + 1),
@@ -74,12 +74,12 @@ test('be_derived_pair_|-base_name|+derived__new|+be__params', ()=>{
 	equal(foobar__([ctx_(), ctx]).$, 3)
 	equal(foobar_([ctx_(), ctx]), 3)
 })
-test('be_derived_pair_|-base_name|+derived__new|-be__params', ()=>{
+test('be_readable_pair_|-base_name|+derived__new|-be__params', ()=>{
 	const base_ = writable_(1)
 	const [
 		foobar__,
 		foobar_,
-	] = be_derived_pair_(
+	] = be_readable_pair_(
 		()=>
 			derived_(base_,
 				base=>base + 1))