+++ src/store/index.d.ts
@@ -1,5 +1,5 @@
 export * from './_types'
-export * from './be_derived_pair_'
+export * from './be_readable_pair_'
 export * from './be_writable_triple_'
 export * from './change_once_subscribe'
 export * from './clear_store'