+++ src/store/index.js
@@ -1,4 +1,4 @@
-export * from './be_derived_pair_/index.js'
+export * from './be_readable_pair_/index.js'
 export * from './be_writable_triple_/index.js'
 export * from './change_once_subscribe/index.js'
 export * from './clear_store/index.js'